requirements.txt
pandas
psycopg2-binary
scikit-learn
mlflow
requests
